split filter to source file
